refactor(engine): split backend-controller into pool modules
backend-controller.ts was 733 lines of pool logic sharing module-level
mutable state across every function. Split into a backend-controller/
folder:
  - types   — public types (roles, holders, snapshot, result)
  - holders — holder-string constructors
  - state   — shared pool/bindings/listeners Maps + internal helpers;
              reassignable initCtx/poolConfig live on a `ctx` holder
              object so other modules can mutate them
  - pool    — init/teardown, role accessors, availability, snapshot
  - rebind  — rebind/release holders + per-chat accessors
  - legacy  — single-active aliases routed to the chat role

index.ts re-exports the identical public surface; all 19 importers
updated to the barrel path. Verified: backend-pool + backend-controller
suites pass (34 tests).
